Seperated gallery from start screen
