+++ Wabbajack/Views/MainWindow.xaml
@@ -22,6 +22,9 @@
                 <DataTemplate DataType="{x:Type local:StartupVM}">
                     <views:StartupView/>
                 </DataTemplate>
+                <DataTemplate DataType="{x:Type local:ModListGalleryVM}">
+                    <views:ModListGalleryView/>
+                </DataTemplate>
                 <!--
             <DataTemplate DataType="{x:Type local:CompilerVM}">
                 <local:CompilerView />
